+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Changed
 - Rust: MSRV is 1.65
+- All held keys are released when Enigo is dropped
 
 ## Added
 - Linux: Support X11 without `xdotools`. Use the experimental feature `x11rb` to test it